+++ release-notes.md
@@ -1,4 +1,4 @@
-## Unreleased
+## NEXT
 * 213008 - BUG accessibility -Prepare - cant create conversion when javascript turned off
 
 ## 1.7.1