+++ release-notes.md
@@ -4,6 +4,8 @@
 
 ## 1.8.0
 * 240477 - Rebrand of head and footer 
+## NEXT
+* 213008 - BUG accessibility -Prepare - cant create conversion when javascript turned off
 
 ## 1.7.1
 * 235754 - Performance data references TRAMS incorrectly - Two other instances